[4.0.2] Bug & Performance Fixes
Users on 4.0 are recommended to update to this version.

- Fixed crash when event gets called multiple times for auto data removal
- Improved performance on FlightDataTask (tested)

+++ src/AGTHARN/FlyPE/EventListener.php
@@ -132,7 +132,7 @@ class EventListener implements Listener {
                 }
                 $playerData->setFlightState(false);
             }
-            if ($playerData->checkNew()) {
+            if (file_exists($playerData->getDataPath()) && $playerData->checkNew()) {
                 unlink($playerData->getDataPath());
             }
             $playerData->saveData();
